+++ src/evm/protocol/ekubo/tycho_decoder.rs
@@ -21,7 +21,10 @@ use super::{
 };
 use crate::{
     evm::protocol::ekubo::pool::mev_resist::MevResistPool,
-    protocol::{errors::InvalidSnapshotError, models::TryFromWithBlock},
+    protocol::{
+        errors::InvalidSnapshotError,
+        models::{DecoderContext, TryFromWithBlock},
+    },
 };
 
 enum EkuboExtension {
@@ -57,6 +60,7 @@ impl TryFromWithBlock<ComponentWithState, BlockHeader> for EkuboState {
         _block: BlockHeader,
         _account_balances: &HashMap<Bytes, HashMap<Bytes, Bytes>>,
         _all_tokens: &HashMap<Bytes, Token>,
+        _decoder_context: &DecoderContext,
     ) -> Result<Self, Self::Error> {
         let static_attrs = snapshot.component.static_attributes;
         let state_attrs = snapshot.state.attributes;
@@ -213,6 +217,7 @@ mod tests {
             BlockHeader::default(),
             &HashMap::new(),
             &HashMap::new(),
+            &DecoderContext::new(),
         )
         .await
         .expect("reconstructing state");
@@ -248,6 +253,7 @@ mod tests {
                 BlockHeader::default(),
                 &HashMap::default(),
                 &HashMap::default(),
+                &DecoderContext::new(),
             )
             .await;