+++ src/evm/protocol/uniswap_v4/state.rs
@@ -822,7 +822,7 @@ mod tests {
             },
             protocol::uniswap_v4::hooks::generic_vm_hook_handler::GenericVMHookHandler,
         },
-        protocol::models::TryFromWithBlock,
+        protocol::models::{DecoderContext, TryFromWithBlock},
     };
 
     // Helper methods to create commonly used tokens
@@ -958,6 +958,7 @@ mod tests {
             block,
             &Default::default(),
             &Default::default(),
+            &DecoderContext::new(),
         )
         .await
         .unwrap();
@@ -1039,6 +1040,7 @@ mod tests {
             block,
             &Default::default(),
             &Default::default(),
+            &DecoderContext::new(),
         )
         .await
         .unwrap();